% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggexample-package.R
\docType{package}
\name{ggexample-package}
\alias{ggexample}
\alias{ggexample-package}
\title{ggexample: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Charlotte Dawson \email{csdaw@outlook.com} (\href{https://orcid.org/0000-0002-7151-5971}{ORCID})

}
\keyword{internal}
